add recipe page, images,rename files to lowercase
Did some renaming of files to uniform lowercase naming convention with hyphens. html will be the root directory for html files and everything else related to html.
I am thinking I will make a folder for each language in the top root directory and push to same repository. This is starting to feel like a mistake but I want to keep everything in one space. so makes sense to me.
Added recipe pages for hot chocolate, sunny side up egg and strawberry jam along with images. shifted previous home page content to an e-log page.
Added a navigation menu to all pages at the top of each page for easy navigation between pages.
Added go to top option at the end of each recipe.
Should I make an assets folder for images and css? or keep it simple for now? Thinking about it...